+++ vibration_platform_interface/lib/vibration_platform_interface.dart
@@ -89,6 +89,7 @@ abstract class VibrationPlatform extends PlatformInterface {
     int repeat = -1,
     List<int> intensities = const [],
     int amplitude = -1,
+    double sharpness = 0.5,
   }) {
     throw UnimplementedError('deviceInfo() has not been implemented.');
   }