+++ vibration/CHANGELOG.md
@@ -1,3 +1,9 @@
+### 3.1.0
+
+- Add common vibration patterns for Android and iOS.
+- Add `sharpness` parameter for iOS.
+- Suppress deprecation warnings for `vibrate` method on Android.
+
 ### 3.0.0
 
 - The plugin has been recreated from scratch to align with the latest Flutter and Dart features.