+++ vibration_platform_interface/CHANGELOG.md
@@ -1,7 +1,11 @@
+## 0.1.0
+
+- Add sharpness parameter for iOS.
+
 ## 0.0.2
 
-* Update package `device_info_plus` to ">=9.0.2 <12.0.0"
+- Update package `device_info_plus` to ">=9.0.2 <12.0.0"
 
 ## 0.0.1
 
-* Initial open-source release.
+- Initial open-source release.